+++ Purchases/Classes/RCPurchases+Protected.h
@@ -14,11 +14,12 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface RCPurchases (Protected)
 
-- (instancetype _Nullable)initWithAppUserID:(NSString *)appUserID
+- (instancetype _Nullable)initWithAppUserID:(NSString * _Nullable)appUserID
                              requestFetcher:(RCStoreKitRequestFetcher *)requestFetcher
                                     backend:(RCBackend *)backend
                             storeKitWrapper:(RCStoreKitWrapper *)storeKitWrapper
-                         notificationCenter:(NSNotificationCenter *)notificationCenter;
+                         notificationCenter:(NSNotificationCenter *)notificationCenter
+                               userDefaults:(NSUserDefaults *)userDefaults;
 
 @end